Move Client code to a source file
